refactor(67): move support settings and placeholder to components
