+++ src/components/NavigationBar/NavigationBar.tsx
@@ -4,9 +4,9 @@ import { NavigationBarProps, TNavigationItem } from './types';
 import { NavigationBarView } from './NavigationBarView';
 import { UserSettingsDialog } from '../UserSettingsDialog';
 import { SettingsSlideIn } from '../SettingsSlideIn';
-import { SettingsSectionItem } from '../../pages/SettingsPage/types';
-import { SupportSettingsPage } from '../../pages/SupportSettingsPage';
-import { SettingsPlaceholder } from '../../pages/SettingsPlaceholder';
+import { SettingsSectionItem } from '../SettingsSlideIn/types';
+import { SupportSettingsPage } from '../SupportSettings';
+import { SettingsPlaceholder } from '../SettingsPlaceholder';
 
 export const NavigationBar = (props: NavigationBarProps) => {
     const { selected, onClick, compact } = props;
