--- src/pages/SettingsPage/index.ts
@@ -1,2 +1,0 @@
-export * from './SettingsPage';
-export * from './types';
