--- src/pages/SettingsPage/types.ts
@@ -1,9 +1,0 @@
-export type SettingsSectionItem = {
-    label: string;
-    onPress: () => void;
-};
-
-export interface SettingsPageViewProps {
-    sections: SettingsSectionItem[];
-    onClose: () => void;
-}
