--- src/pages/SupportSettingsPage/SupportSettingsPage.tsx
@@ -1,56 +1,0 @@
-import React, { useState, useEffect, useRef, useCallback } from 'react';
-import { Share, Linking } from 'react-native';
-import { SupportSettingsDisplayProps, useSupportSettingsDisplay } from 'incyclist-services';
-import { useLogging, useUnmountEffect } from '../../hooks';
-import { SupportSettingsView } from './View';
-
-interface SupportSettingsPageProps {
-    onClose: () => void;
-}
-
-export const SupportSettingsPage = ({ onClose }: SupportSettingsPageProps) => {
-    const [displayProps, setDisplayProps] = useState<SupportSettingsDisplayProps | null>(null);
-    const refInitialized = useRef(false);
-    const service = useSupportSettingsDisplay();
-    const { logEvent } = useLogging('SupportSettingsPage');
-
-    useEffect(() => {
-        if (refInitialized.current) return;
-        service.open();
-        setDisplayProps(service.getDisplayProps());
-        refInitialized.current = true;
-    }, [service]);
-
-    useUnmountEffect(() => {
-        service.close();
-    });
-
-    const handleClose = useCallback(() => {
-        logEvent({ message: 'button clicked', button: 'back', eventSource: 'user' });
-        onClose();
-    }, [onClose, logEvent]);
-
-    const onShareUuid = useCallback(async () => {
-        if (!displayProps) return;
-        logEvent({ message: 'button clicked', button: 'share uuid', eventSource: 'user' });
-        try {
-            await Share.share({ message: displayProps.uuid });
-        } catch (err: any) {
-            logEvent({ message: 'sharing cancelled or failed', reason: err.message });
-        }
-    }, [displayProps, logEvent]);
-
-    const onOpenUrl = useCallback((url: string) => {
-        logEvent({ message: 'link clicked', url, eventSource: 'user' });
-        Linking.openURL(url);
-    }, [logEvent]);
-
-    return (
-        <SupportSettingsView
-            displayProps={displayProps}
-            onClose={handleClose}
-            onShareUuid={onShareUuid}
-            onOpenUrl={onOpenUrl}
-        />
-    );
-};
