--- src/pages/SupportSettingsPage/types.ts
@@ -1,8 +1,0 @@
-import { SupportSettingsDisplayProps } from 'incyclist-services';
-
-export interface SupportSettingsViewProps {
-    displayProps: SupportSettingsDisplayProps | null;
-    onClose: () => void;
-    onShareUuid: () => void;
-    onOpenUrl: (url: string) => void;
-}
